refactor forcing zarr access to a separate file
