+++ modules/data_processing/forcings.py
@@ -4,7 +4,6 @@ import os
 import time
 import warnings
 from pathlib import Path
-from typing import Tuple
 from datetime import datetime
 from functools import partial
 
@@ -14,12 +13,12 @@ import dask
 from dask.distributed import Client, LocalCluster
 import geopandas as gpd
 import pandas as pd
-import s3fs
 import xarray as xr
 from exactextract import exact_extract
 from multiprocessing import shared_memory
 
 from data_processing.file_paths import file_paths
+from data_processing.zarr_utils import get_forcing_data
 
 logger = logging.getLogger(__name__)
 # Suppress the specific warning from numpy
@@ -31,51 +30,6 @@ warnings.filterwarnings(
 )
 
 
-def open_s3_store(url: str) -> s3fs.S3Map:
-    """Open an s3 store from a given url."""
-    return s3fs.S3Map(url, s3=s3fs.S3FileSystem(anon=True))
-
-
-def load_zarr_datasets() -> xr.Dataset:
-    """Load zarr datasets from S3 within the specified time range."""
-    # if a LocalCluster is not already running, start one
-    if not Client(timeout="2s"):
-        cluster = LocalCluster()
-    forcing_vars = ["lwdown", "precip", "psfc", "q2d", "swdown", "t2d", "u2d", "v2d"]
-    s3_urls = [
-        f"s3://noaa-nwm-retrospective-3-0-pds/CONUS/zarr/forcing/{var}.zarr"
-        for var in forcing_vars
-    ]
-    s3_stores = [open_s3_store(url) for url in s3_urls]
-    dataset = xr.open_mfdataset(s3_stores, parallel=True, engine="zarr")
-    return dataset
-
-
-def load_geodataframe(geopackage_path: str, projection: str) -> gpd.GeoDataFrame:
-    """Load and project a geodataframe from a given path and projection."""
-    gdf = gpd.read_file(geopackage_path, layer="divides").to_crs(projection)
-    return gdf
-
-
-def clip_dataset_to_bounds(
-    dataset: xr.Dataset, bounds: Tuple[float, float, float, float], start_time: str, end_time: str
-) -> xr.Dataset:
-    """Clip the dataset to specified geographical bounds."""
-    dataset = dataset.sel(
-        x=slice(bounds[0], bounds[2]),
-        y=slice(bounds[1], bounds[3]),
-        time=slice(start_time, end_time),
-    )
-    logger.info("Selected time range and clipped to bounds")
-    return dataset
-
-
-def compute_store(stores: xr.Dataset, cached_nc_path: Path) -> xr.Dataset:
-    stores.to_netcdf(cached_nc_path)
-    data = xr.open_mfdataset(cached_nc_path, parallel=True, engine="h5netcdf")
-    return data
-
-
 def weighted_sum_of_cells(flat_tensor, cell_ids, factors):
     # Create an output array initialized with zeros
     result = np.zeros(flat_tensor.shape[0])
@@ -83,7 +37,6 @@ def weighted_sum_of_cells(flat_tensor, cell_ids, factors):
     sum_of_weights = np.sum(factors)
     result /= sum_of_weights
     return result
-
 
 def get_cell_weights(raster, gdf):
     output = exact_extract(
@@ -241,7 +194,6 @@ def compute_zonal_stats(
 
 def setup_directories(wb_id: str) -> file_paths:
     forcing_paths = file_paths(wb_id)
-
     for folder in ["by_catchment", "temp"]:
         os.makedirs(forcing_paths.forcings_dir() / folder, exist_ok=True)
     return forcing_paths
@@ -252,7 +204,7 @@ def create_forcings(start_time: str, end_time: str, output_folder_name: str) -> 
     projection = xr.open_dataset(forcing_paths.template_nc(), engine="h5netcdf").crs.esri_pe_string
     logger.debug("Got projection from grid file")
 
-    gdf = load_geodataframe(forcing_paths.geopackage_path(), projection)
+    gdf = gpd.read_file(forcing_paths.geopackage_path(), layer="divides").to_crs(projection)
     logger.debug("Got gdf")
 
     if type(start_time) == datetime:
@@ -260,39 +212,7 @@ def create_forcings(start_time: str, end_time: str, output_folder_name: str) -> 
     if type(end_time) == datetime:
         end_time = end_time.strftime("%Y-%m-%d %H:%M")
 
-    merged_data = None
-    if os.path.exists(forcing_paths.cached_nc_file()):
-        logger.info("Found cached nc file")
-        # open the cached file and check that the time range is correct
-        cached_data = xr.open_mfdataset(
-            forcing_paths.cached_nc_file(), parallel=True, engine="h5netcdf"
-        )
-        if cached_data.time[0].values <= np.datetime64(start_time) and cached_data.time[
-            -1
-        ].values >= np.datetime64(end_time):
-            logger.info("Time range is correct")
-            logger.debug(f"Opened cached nc file: [{forcing_paths.cached_nc_file()}]")
-            merged_data = clip_dataset_to_bounds(
-                cached_data, gdf.total_bounds, start_time, end_time
-            )
-            logger.debug("Clipped stores")
-        else:
-            logger.info("Time range is incorrect")
-            os.remove(forcing_paths.cached_nc_file())
-            logger.debug("Removed cached nc file")
-
-    if merged_data is None:
-        logger.info("Loading zarr stores, this may take a while.")
-        lazy_store = load_zarr_datasets()
-        logger.debug("Got zarr stores")
-
-        clipped_store = clip_dataset_to_bounds(lazy_store, gdf.total_bounds, start_time, end_time)
-        logger.info("Clipped forcing data to bounds")
-
-        merged_data = compute_store(clipped_store, forcing_paths.cached_nc_file())
-        logger.info("Forcing data loaded and cached")
-
-    logger.info("Computing zonal stats")
+    merged_data = get_forcing_data(forcing_paths, start_time, end_time, gdf)
     compute_zonal_stats(gdf, merged_data, forcing_paths.forcings_dir())
